Use module attribute `__file__` instead of string
The `base_path` was using the string `"__file__"` instead of the module
attribute `__file__` (no quotes). This was probably the reason for the
`joinpath` hack used to create the correct base path.

See: https://docs.python.org/3/reference/datamodel.html#index-42

The problem with using the string `"__file__"` is that will create a
`Path` pointing to an actual file/dir named `"__file__"` in the working
directory. Since we typically run the bot with the repository root as
the working directory, this necessitated the `joinpath` hack with
 "EuroPythonBot"` to recreate a path pointing to the actual `base_dir`.

By using `__file__` (no quotes), the base_dir is determined relative to
`configuration.py`, which seems to have been the original intention
behind the code, instead of relative to whatever happens to be the
working directory.