Remove default values in registration interaction
The default values were confusing and obscured the placeholder text with
more information about the fields. Not having default values also means
that participants can't hit enter until they actually supplied a value
of their own.

A lot of the current registration errors in the log show that people
press enter with the default values still in place.