Add user id to registration logs
By logging the user id, we create a relationship between a Discord
account and a registration using an immutable reference. The other
references that we log, username and nickname, are both editable by the
user.

Having an immutable relationship is important for situations in which we
need to be able to determine identity, like Code of Conduct enforcement.